Move ConjureProductDependenciesExtension to api package
